finish design. update code

+++ client/src/components/EditTask.js
@@ -21,11 +21,13 @@ class EditTask extends React.Component {
     };
 
     handleDeleteTask = () => {
-        fetch(`/api/tasks/${this.props.task.id}`, {
-            method: 'DELETE'
-        })
-            .then((response) => this.props.updateList())
-            .catch((error) => console.error('Error:', error));
+        if (window.confirm('Are you sure you want to delete this task?')) {
+            fetch(`/api/tasks/${this.props.task.id}`, {
+                method: 'DELETE'
+            })
+                .then((response) => this.props.updateList())
+                .catch((error) => console.error('Error:', error));
+        }
     };
 
     render() {
@@ -35,16 +37,16 @@ class EditTask extends React.Component {
                     onChange={(e) => this.setState({ text: e.target.value })} />
                     : null}
 
-                {this.state.editing ? <button onClick={this.handleUpdateTask}>Update</button>
+                {this.state.editing
+                    ? <button onClick={this.handleUpdateTask}>Done</button>
                     : <div>
                         <button onClick={() => this.setState({ editing: true }/* enable editing mode */)}>
-                            <i className="fas fa-pencil-alt"></i>
+                            <i className="fa fa-pencil-alt fa-sm"></i>
                         </button>
                         <button onClick={this.handleDeleteTask}>
-                            <i className="fas fa-trash-alt"></i>
+                            <i className="fa fa-trash-alt fa-sm"></i>
                         </button>
-                    </div>
-                }
+                    </div>}
             </div>
         );
     }
